1.03
git-svn-id: http://svn.vuze.com/public/client@10647 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/platform/macosx/access/jnilib/OSXAccess.c
@@ -117,7 +117,7 @@ Java_org_gudy_azureus2_platform_macosx_access_jnilib_OSXAccess_memmove(
 	jint src,
 	jint count)
 {
-	jshort *dest1;
+	jbyte *dest1;
 	
 	if (dest) {
 		dest1 = (*env)->GetByteArrayElements(env, dest, NULL);
